Creating the DatabaseSQL class, and moving all the SQLite functions from Database to DatabaseSQL.
