+++ src/frontend/model_settings.yaml
@@ -1,9 +1,13 @@
+
 model_name: test_case_1
 model_description: dummy case study for testing the model
 model_data_folder_path: D:\Politecnico di Milano\Documentale DENG - PRIN-MIMO\Models\pyESM_docs\case_studies
+
 database_settings:
-    generate_blank_sets: false
+    generate_blank_sets: true
     sets_file_name: sets.xlsx
-    sets_definition_excel: true
+    generate_database_sql: true
+    database_sql_name: database.db
+
 problem_settings:
     execution_mode: simulation
